# backend/api/public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentação de API da aplicação Intellecta",
        "version": "0.1"
    },
    "paths": {
...
        "/institutions/{institution_id}/subjects": {
            "get": {
                "tags": [
                    "Subjects"
                ],
                "summary": "Get all subjects for an institution",
                "operationId": "getInstitutionSubjects",
                "parameters": [
                    {
                        "name": "institution_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of subjects",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SubjectResponse"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Institution not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Subjects"
                ],
                "summary": "Create a new subject",
                "description": "Create a new subject in the specified institution. Admins can create subjects for any teacher, teachers can only create subjects for themselves.",
                "operationId": "createSubject",
                "parameters": [
                    {
                        "name": "institution_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Subject creation data",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateSubjectRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Subject created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "subject": {
                                            "$ref": "#/components/schemas/SubjectResponse"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "403": {
                        "description": "Forbidden - only admins and teachers can create subjects"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/institutions/{institution_id}/subjects/{subject_id}": {
            "get": {
                "tags": [
                    "Subjects"
                ],
                "summary": "Get subject details",
                "description": "Retrieve details of a specific subject. Only accessible by admins and the subject teacher.",
                "operationId": "getSubjectById",
                "parameters": [
                    {
                        "name": "institution_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "subject_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Subject details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SubjectResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - not admin or subject teacher"
                    },
                    "404": {
                        "description": "Subject not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Subjects"
                ],
                "summary": "Update a subject",
                "description": "Update details of an existing subject. Only admins can change the teacher.",
                "operationId": "updateSubjectById",
                "parameters": [
                    {
                        "name": "institution_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "subject_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Subject update data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateSubjectRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated subject details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SubjectResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "403": {
                        "description": "Forbidden - not admin or subject teacher"
                    },
                    "404": {
                        "description": "Subject or file not found"
                    },
                    "422": {
                        "description": "Invalid file type (must be image)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Subjects"
                ],
                "summary": "Delete a subject",
                "description": "Permanently delete a subject. Only accessible by admins and the subject teacher.",
                "operationId": "deleteSubjectById",
                "parameters": [
                    {
                        "name": "institution_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "subject_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Subject deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - not admin or subject teacher"
                    },
                    "404": {
                        "description": "Subject not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "SubjectResponse": {
                "properties": {
                    "subject_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "institution_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "teacher_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "profile_picture_id": {
                        "type": "string",
                        "format": "uuid",
                        "nullable": true
                    },
                    "banner_id": {
                        "type": "string",
                        "format": "uuid",
                        "nullable": true
                    },
                    "teacher": {
                        "$ref": "#/components/schemas/UserResponse"
                    },
                    "profile_picture": {
                        "type": "string",
                        "nullable": true
                    },
                    "banner": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "CreateSubjectRequest": {
                "required": [
                    "name",
                    "description"
                ],
                "properties": {
                    "name": {
                        "description": "Name of the subject",
                        "type": "string",
                        "maxLength": 255,
                        "minLength": 3
                    },
                    "description": {
                        "description": "Detailed description of the subject",
                        "type": "string"
                    },
                    "teacher_id": {
                        "description": "ID of the teacher (required if admin, optional for teachers)",
                        "type": "string",
                        "format": "uuid",
                        "nullable": true
                    },
                    "profile-picture": {
                        "description": "Profile picture file",
                        "type": "string",
                        "format": "binary",
                        "nullable": true
                    },
                    "banner": {
                        "description": "Banner image file",
                        "type": "string",
                        "format": "binary",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "UpdateSubjectRequest": {
                "required": [
                    "name",
                    "description"
                ],
                "properties": {
                    "name": {
                        "description": "Updated name of the subject",
                        "type": "string",
                        "maxLength": 255,
                        "minLength": 3
                    },
                    "description": {
                        "description": "Updated description of the subject",
                        "type": "string"
                    },
                    "teacher_id": {
                        "description": "Updated teacher ID (admin only)",
                        "type": "string",
                        "format": "uuid",
                        "nullable": true
                    },
                    "profile_picture_id": {
                        "description": "ID of the new profile picture",
                        "type": "string",
                        "format": "uuid",
                        "nullable": true
                    },
                    "banner_id": {
                        "description": "ID of the new banner image",
                        "type": "string",
                        "format": "uuid",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "UserResponse": {
                "properties": {
                    "user_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "full_name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "profile_picture": {
                        "properties": {
                            "file_id": {
                                "type": "string",
                                "format": "uuid"
                            },
                            "url": {
                                "type": "string",
                                "format": "uri"
                            },
                            "filename": {
                                "type": "string"
                            },
                            "mime_type": {
                                "type": "string"
                            },
                            "size": {
                                "type": "integer"
                            }
                        },
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}